% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chng_pnt_algo.R
\name{chng_pnt_algo}
\alias{chng_pnt_algo}
\title{chng_pnt_algo}
\usage{
chng_pnt_algo(
  v,
  dat,
  namX = "X",
  namY = "Y",
  col_pt = rgb(0, 0, 0, 0.125),
  col_app = "red",
  col_op = "blue",
  gr_app = T,
  gr_op = T,
  gr_leg=T
)
}
\arguments{
\item{v}{The vector which is the concatenation of the (approximate/estimated/initial guess)breakpoints and the (estimated/initial guess)plateau heights vectors}

\item{dat}{The two column dataframe over which you want to optimize the (plateau/slope) curve. The first column is the x-axis, while the second column is the y-axis}

\item{namX}{The X-axis label of the output graph(default to 'X')}

\item{namY}{The Y-axis label of the output graph(default to 'Y')}

\item{col_pt}{The datapoints colour of the output scatter plot(recommended to use rgb(R,G,B,Tr)) of your raw data}

\item{col_app}{The colour of the approximated (estimated/initial guess) lines/curve overlayed on your raw data plot(dat), generated using your input vector(v)}

\item{col_op}{The colour of the calculated/optimized(calculated/optimized by the chng_pnt_algo) line/curve overlayed over your raw data(dat)}

\item{gr_app}{A binary bit(default set to T) asking whether the user wants to see the approximated/estimated/initial guess curve, derived from vector v, overlayed on the raw data(dat)}

\item{gr_op}{A binary bit(default set to T) asking whether the user wants to see the optimized/calculated line/curve overlayed on the raw data plot(data)}
\item{gr_leg}{A binary bit(default set to T) asking whether the user wants to see the legend over the plot}
}
\value{
Returns a list L, containing information about the optimization calculation with the new optimized vector v , new breakpoints Bp and the new heights H.
}
\description{
chng_pnt_algo
}
