\name{num_diff}
\alias{num_diff}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Returns a dataframe consisting of the numerically differentiated second variable w.r.t the first variable (dY/dX) as the second variable and (X) as the first(independent variable). It also plots the curve depending on the input arguments
}
\description{
The function allows for the numeric differentiation of the second column of your data frame with respect to the first
}
\usage{
num_diff(dframe, pl=T)
}
\arguments{
  \item{dframe}{
The input 2-column dataframe with the independent variable(X) on the first column and the dependent variable(Y) on the second column.
}
  \item{pl}{
Set to true by default, the argument can be toggled in the light of necessity, to deal with whether you want the function to plot the output dataframe or not. pl=F implies no output plot. It is set to True by default.
  }
}
\details{
The function always returns a dataframe. There are no other return types or no way to tune the return type in this function.
}
\value{
The return type of the function is always a 2-column dataframe with the 1st column being the independent variable, while the next column is the dependent variable(dY/dX)
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Chitran Ghosal
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
df <- read.table('X:/Dido/Drive/Didos_self/Career/Future/PhD/Trash/test.txt')
num_diff(df, pl = F)

> df <- read.table('X:/Dido/Drive/Didos_self/Career/Future/PhD/Trash/test.txt')
> num_diff(df, pl = F)
              X             Y
1   -0.99497500  1.264054e-09
2   -0.98492500  1.430204e-09
3   -0.97487500  1.435568e-09
4   -0.96482500  1.269268e-09
5   -0.95477500  1.625556e-09
6   -0.94472500  1.354463e-09
7   -0.93467500  1.551659e-09
8   -0.92462500  1.727597e-09
9   -0.91457500  1.688713e-09
10  -0.90452500  1.481421e-09
11  -0.89447500  1.722762e-09
12  -0.88442500  1.567588e-09
13  -0.87437500  1.170320e-09
14  -0.86432500  1.272423e-09
15  -0.85427500  1.288047e-09
16  -0.84422500  1.271796e-09
17  -0.83417500  1.386543e-09
18  -0.82412500  9.421534e-10
19  -0.81407500  8.995536e-10
20  -0.80402000  1.185831e-09
21  -0.79396500  1.122439e-09
22  -0.78391500  9.223151e-10
23  -0.77386500  7.942294e-10
24  -0.76381500  1.129101e-09
25  -0.75376500  1.091319e-09
26  -0.74371500  1.029001e-09
27  -0.73366500  8.763273e-10
28  -0.72361500  1.196328e-09
29  -0.71356500  1.027493e-09
30  -0.70351500  8.913821e-10
31  -0.69346500  1.039035e-09
32  -0.68341500  1.223298e-09
33  -0.67336500  1.078022e-09
34  -0.66331500  1.110877e-09
35  -0.65326500  1.117524e-09
36  -0.64321500  7.836385e-10
37  -0.63316500  1.261341e-09
38  -0.62311500  1.247849e-09
39  -0.61306500  1.294306e-09
40  -0.60301500  1.290429e-09
41  -0.59296500  1.449700e-09
42  -0.58291500  1.590412e-09
43  -0.57286500  1.565193e-09
44  -0.56281500  1.568105e-09
45  -0.55276500  1.652374e-09
46  -0.54271500  1.724338e-09
47  -0.53266500  1.954530e-09
48  -0.52261500  1.982910e-09
49  -0.51256500  2.129029e-09
50  -0.50251500  2.143860e-09
51  -0.49246500  2.355489e-09
52  -0.48241500  2.406216e-09
53  -0.47236500  2.164965e-09
54  -0.46231500  2.269116e-09
55  -0.45226500  2.307505e-09
56  -0.44221500  2.036449e-09
57  -0.43216500  1.918176e-09
58  -0.42211500  1.707480e-09
59  -0.41206500  1.507721e-09
60  -0.40201000  1.337250e-09
61  -0.39195500  1.139874e-09
62  -0.38190500  1.105189e-09
63  -0.37185500  9.679549e-10
64  -0.36180500  8.824264e-10
65  -0.35175500  7.935711e-10
66  -0.34170500  7.497606e-10
67  -0.33165500  6.809341e-10
68  -0.32160500  6.220258e-10
69  -0.31155500  5.362371e-10
70  -0.30150500  4.992687e-10
71  -0.29145500  4.969681e-10
72  -0.28140500  4.737544e-10
73  -0.27135500  4.385856e-10
74  -0.26130500  3.852075e-10
75  -0.25125500  3.184230e-10
76  -0.24120500  2.522408e-10
77  -0.23115500  1.752920e-10
78  -0.22110500  1.107858e-10
79  -0.21105500  6.498465e-11
80  -0.20100500  4.514808e-11
81  -0.19095500  3.250966e-11
82  -0.18090500  2.185112e-11
83  -0.17085500  1.715156e-11
84  -0.16080500  1.044408e-11
85  -0.15075500 -4.148872e-13
86  -0.14070500 -1.016974e-11
87  -0.13065500 -1.261056e-11
88  -0.12060500 -1.123684e-11
89  -0.11055500 -9.051797e-12
90  -0.10050350  3.502593e-12
91  -0.09045200  8.632353e-12
92  -0.08040200  1.098673e-11
93  -0.07035200  1.154705e-11
94  -0.06030150  1.355226e-11
95  -0.05025100  2.441902e-11
96  -0.04020100  5.440420e-11
97  -0.03015100  8.840664e-11
98  -0.02010050  1.164013e-10
99  -0.01005005  1.665064e-10
100  0.00000000  2.574463e-10
101  0.01005005  3.628937e-10
102  0.02010050  4.790655e-10
103  0.03015100  5.679360e-10
104  0.04020100  6.560388e-10
105  0.05025100  7.470719e-10
106  0.06030150  8.749222e-10
107  0.07035200  9.894197e-10
108  0.08040200  1.104592e-09
109  0.09045200  1.207807e-09
110  0.10050350  1.400875e-09
111  0.11055500  1.594230e-09
112  0.12060500  1.743042e-09
113  0.13065500  2.009145e-09
114  0.14070500  2.254723e-09
115  0.15075500  2.462396e-09
116  0.16080500  2.674815e-09
117  0.17085500  2.877410e-09
118  0.18090500  3.175769e-09
119  0.19095500  3.357153e-09
120  0.20100500  3.330708e-09
121  0.21105500  3.712523e-09
122  0.22110500  3.559733e-09
123  0.23115500  3.409623e-09
124  0.24120500  3.537301e-09
125  0.25125500  3.191040e-09
126  0.26130500  2.711186e-09
127  0.27135500  3.059638e-09
128  0.28140500  2.600263e-09
129  0.29145500  2.416624e-09
130  0.30150500  2.316141e-09
131  0.31155500  2.280756e-09
132  0.32160500  2.388435e-09
133  0.33165500  2.138479e-09
134  0.34170500  2.352421e-09
135  0.35175500  2.139983e-09
136  0.36180500  2.126904e-09
137  0.37185500  2.337509e-09
138  0.38190500  1.787904e-09
139  0.39195500  1.729674e-09
140  0.40201000  1.936741e-09
141  0.41206500  1.475262e-09
142  0.42211500  1.265484e-09
143  0.43216500  1.285780e-09
144  0.44221500  9.888956e-10
145  0.45226500  1.042953e-09
146  0.46231500  5.430339e-10
147  0.47236500  4.199545e-10
148  0.48241500  6.485980e-10
149  0.49246500  5.091826e-11
150  0.50251500  5.863063e-10
151  0.51256500 -2.333664e-10
152  0.52261500 -8.628732e-11
153  0.53266500 -2.583920e-10
154  0.54271500 -4.278819e-10
155  0.55276500 -3.001785e-10
156  0.56281500 -5.946338e-10
157  0.57286500 -4.259357e-10
158  0.58291500 -3.389922e-10
159  0.59296500 -2.182646e-10
160  0.60301500 -4.204376e-10
161  0.61306500 -1.755399e-10
162  0.62311500 -9.308750e-11
163  0.63316500  3.540710e-10
164  0.64321500  9.335884e-11
165  0.65326500  3.918780e-10
166  0.66331500  6.130071e-10
167  0.67336500  1.083296e-09
168  0.68341500  9.423212e-10
169  0.69346500  1.426604e-09
170  0.70351500  1.542052e-09
171  0.71356500  1.715193e-09
172  0.72361500  2.289820e-09
173  0.73366500  2.267392e-09
174  0.74371500  2.107115e-09
175  0.75376500  2.923075e-09
176  0.76381500  3.188492e-09
177  0.77386500  3.335144e-09
178  0.78391500  3.273904e-09
179  0.79396500  3.542950e-09
180  0.80402000  3.659189e-09
181  0.81407500  3.906818e-09
182  0.82412500  3.567899e-09
183  0.83417500  3.683923e-09
184  0.84422500  3.972388e-09
185  0.85427500  3.340866e-09
186  0.86432500  3.927235e-09
187  0.87437500  3.839755e-09
188  0.88442500  2.832235e-09
189  0.89447500  2.744636e-09
190  0.90452500  3.441687e-09
191  0.91457500  3.067611e-09
192  0.92462500  1.957740e-09
193  0.93467500  2.284632e-09
194  0.94472500  2.828916e-09
195  0.95477500  1.227306e-09
196  0.96482500  1.736894e-09
197  0.97487500  1.897047e-09
198  0.98492500  8.831217e-10
199  0.99497500  1.234971e-09




num_diff <- function(dframe, pl=T){
  x <-dframe$V1
  y <-dframe$V2
  x1 <- NA
  y1 <- NA
  for (i in 1:(length(x)-1)) {
    x1 <- c(x1,(x[i+1]+x[i])/2)
    y1 <- c(y1,(y[i+1]-y[i])/(x[i+1]-x[i]))
  }
  x2 <- x1[!is.na(x1)]
  y2 <- y1[!is.na(y1)]

  if(pl==T){
    plot(x2,y2, pch=19, col=rgb(0,0,1,0.3))
  }
  df_res <- data.frame(matrix(c(x2,y2), ncol = 2, byrow = F))
  names(df_res) <- c('X','Y')
  rm(pl)
  return(df_res)
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ numerical differentiation}% use one of  RShowDoc("KEYWORDS")
\keyword{ numerical calculus}% __ONLY ONE__ keyword per line
