% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsChitran-package.R
\docType{package}
\name{StatsChitran-package}
\alias{StatsChitran}
\alias{StatsChitran-package}
\title{StatsChitran}
\description{
Functions for basic statistical tasks with optional C++ backends via Rcpp.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Chitran1987/StatsChitran}
  \item Report bugs at \url{https://github.com/Chitran1987/StatsChitran/issues}
}

}
\author{
\strong{Maintainer}: Chitran Ghosal \email{ghosal.chitran@gmail.com}

}
\keyword{internal}
