% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsChitran-package.R
\docType{package}
\name{StatsChitran}
\alias{StatsChitran-package}
\alias{StatsChitran}
\title{StatsChitran: Utilities for \if{html}{\out{<your short tagline>}}}
\description{
Package-level documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Chitran1987/StatsChitran}
  \item Report bugs at \url{https://github.com/Chitran1987/StatsChitran/issues}
}

}
\author{
\strong{Maintainer}: Chitran Ghosal \email{ghosal.chitran@gmail.com}

Authors:
\itemize{
  \item Chitran Ghosal \email{ghosal.chitran@gmail.com}
}

}
