% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{function for calculating the linear regression}
\usage{
linreg(df)
}
\arguments{
\item{df}{input a dataframe with more than two columns.\cr
\cr
The last column is always presumed to be the
column of the independent variables in the predictor equation while the the others are presumed to be
the dependent variables.\cr
\cr
The first column of the dataframe is the first regressor variable/independent variable.\cr
\cr
The next column, the second regressor variable/independent variable..so on and so forth till the second last column.\cr
\cr
This is important since the output vector corresponds accordingly.
}
}
\value{
A return vector comprising of your the estimated values of the regression weights (if the input data does not suffer from multicollinearity)
\item{1st_element}{The first element of the vector contains the co-efficient for the first regressor/independent varaiable}
\item{2nd_element}{The second element of the vector contains the co-efficient for the second regressor/independent varaiable so on and so forth till the second last element}
\item{nth_element}{The last element of the returned vector contains the constant value of the predicted equation
}
}
\description{
function for calculating the linear regression
}
\examples{
linreg(reg_data1)
[,1]
[1,] 2.950339
[2,] 5.052787
[3,] 0.999999
[4,] 9.599983
(The reg_data1 dataset is available with the "StatsChitran" package)

linreg(reg_data2)
NULL
(The reg_data2 dataset is available with "StatsChitran" package. The function returns NULL in case
is an issue with multicollinearity in the input dataset)
}
