% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{program for calculating the linear regression by setting the gradients of the regression weights to
zero in the residual sum of squares equation.}
\usage{
linreg(df)
}
\arguments{
\item{df}{input a dataframe with more than two columns. The last column is always presumed to be the
column of the independent variables in the perictor equation while the the others are presumed to be
the dependent variables.}
}
\value{
A dataframe comprising of your the estimated values of the regression weights
}
\description{
program for calculating the linear regression by setting the gradients of the regression weights to
zero in the residual sum of squares equation.
}
